move xml_to_stock_report_helper to commtrack.parsing
